Start kernel stack at end of first page
Instead of address 1000. That was a typo. This gives just over 3kB to
work with for a single syscall, which ought to be more than enough.

This address was chosen because it fits into an immediate value, thus
taking up only instruction and one cycle to execute, and the space being
used is not going to be allocated for any task memory space.

Blowing the stack always has consequences, and in this case the
consequence would be overwriting the interrupt vector and hitting address
0, which would actually cause the hardware to shut off on x86. Not sure
what would happen on ARM.