Rewrite scheduling and tasks to be less memory intensive
This involved changing the data structures, and hence, the logic was
also changed somewhat.

At the same time, data is better grouped so cache performance should be
nicer later on. Also task_create was simplified as the parent id of the
task creator will always be the active task (except for bootstrapping,
which we handle specially).