Rework message passing argument validation to delay a check
This way, we only need to perform the tid and sp checks each time and
can delay the < 0 check on the tid until there is an actual problem.

Suppose you fuck up and pass -1 as the tid for a Send() call. What will
happen is that it will apply the tid mask to -1, which yields 63 in our
current implementation. Then we lookup descriptor 63 and compare the
descriptor tid against the request tid. Since the descriptor tid will
not be negative, the check will fail and we will go into error handling
mode. At that point, we can perform the < 0 check against the request
and return the correct error code.