Handle Delay and DelayUntil errors more intelligently
DelayUntil the past will log a message and wake up right away once it
gets into the server. DelayUntil with a negative value will ABORT because
it implies a very wrong calculation somewhere and not a slightly missed
deadline.

Delay with a negative value will print a log message and return without
even making the system call.