New position tracking for train_blaster
Pretty good at handling basic acceleration cases now. Stuff is in place
to handle checkpointing in between sensors, but it is not yet implemented.

The only problem now is that sometimes starting/stopping after the train
gets lost will cause prediction of the next position to be an exit point,
which causes weird behaviour, possibly keeping the train lost.