Fix some bugs in position tracking
Turns out I was doing reverse incorrectly, which screwed up a lot of
things.

Also turns out I was missing some cases where acceleration data needed
to be merged into the truth and actual sensor data was not available or
out of date which screwed up things a lot.

There still seems to be one issue with position tracking when a sensor
is missed/skipped during acceleration.