Handle reverse at the start of a path correctly
If reverse is the first step, then the second step will be a sensor that
is the reverse of the current sensor, which is not the next_location sensor,
which will cause the path finding logic to screw up.

The solution here is to just handle the initial reverse specially, and 
avoid fast forwarding until the next sensor hit.

Since we only allow for the reverse at start when the train is stopped,
we do not have to worry about the train travelling to another sensor
before it adjusts things.